following the company coding standard

+++ src/main/java/com/hydrogenhr/model/dto/InvoiceDTO.java
@@ -1,6 +1,6 @@
 package com.hydrogenhr.model.dto;
 
-import java.math.BigDecimal;
+
 import java.util.UUID;
 
 import lombok.Getter;
@@ -9,6 +9,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class InvoiceDTO {
-    private BigDecimal amountDue;
+    private Double amountDue;
     private UUID revenueSetupId;
 }
